MILESTONE 1 PUSH
For the first milestone, we completed all of our original requirements. However, we were unable to deliver quality testing. We are still getting used to the testing packages and were having particular difficulty with widget tests. We are not sure what percentage coverage we have for the testing. As for contributions, we feel we have all meaningfully contributed to the project for the first sprint and that we are all doing our part.