MILESTONE 2 PUSH
For the second milestone, we completed all original requirements except for the game timer, and the location leaderboard. The timer was meant to make the numbers disappear after a set time so that the player didn’t have infinite time to memorize the pattern, and would get smaller over time. Julian will complete it for milestone #3. As for the location leaderboard, we have decided that it will just be a global one, as we are already using the photo gallery for the device feature requirement. As for our testing, we have reached around 50% coverage and feel satisfied with our progress there. When it comes to contributions, I feel we have all meaningfully contributed to milestone #2, although I feel Albert and Hansen went above and beyond what I, Julian, was able to accomplish for this milestone. Nonetheless, we all did our part and worked together.