MILESTONE 3 PUSH
For the third milestone, we have completed the goals we set out. Originally we imagined a few more animations, but we still met our requirements and added lots of visuals that make the game look better, as well as updating gameplay with a timer and improving our testing, and UI fixing for hard difficulty. We are proud of our 63% test coverage.